Global content support for search results list
