+++ blocks/search-results-list-block/features/search-results-list/_children/helpers.jsx
@@ -1,0 +1,14 @@
+import getProperties from 'fusion:properties';
+
+const constructHref = (websiteUrl, arcSite) => {
+  const {
+    websiteDomain,
+  } = getProperties(arcSite);
+  return (typeof window !== 'undefined' && window.location.hostname === 'localhost')
+    ? `https://corecomponents-the-gazette-prod.cdn.arcpublishing.com/${websiteUrl}`
+    : `${websiteDomain}/${websiteUrl}`;
+};
+
+const extractImage = (promo) => promo && promo.basic && promo.basic.type === 'image' && promo.basic.url;
+
+export { constructHref, extractImage };
