+++ blocks/search-results-list-block/features/search-results-list/_children/search-results-list.scss
