+++ blocks/search-results-list-block/features/search-results-list/_children/styled-components.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const HeadlineText = styled.h2`
+  font-family: ${(props) => props.primaryFont};
+`;
+
+const DescriptionText = styled.p`
+  font-family: ${(props) => props.secondaryFont};
+`;
+
+export { HeadlineText, DescriptionText };
